generate/2025/2025-09-11_背景抜き/requirements.txt
rembg[new]>=2.0.50
Pillow>=10.0.0